Moved legacy NetworkInterfaces()-related tests to interfaces_test.go
