+++ provider/maas/interfaces_test.go
@@ -4,9 +4,18 @@
 package maas
 
 import (
+	"bytes"
+	"text/template"
+
 	jc "github.com/juju/testing/checkers"
+	"github.com/juju/utils/set"
 	gc "gopkg.in/check.v1"
+
+	"github.com/juju/juju/network"
 )
+
+////////////////////////////////////////////////////////////////////////////////
+// New (1.9 and later) environs.NetworkInterfaces() implementation tests follow.
 
 type interfacesSuite struct {
 	providerSuite
@@ -392,4 +401,368 @@ func (s *interfacesSuite) TestParseInterfacesExampleJSON(c *gc.C) {
 	result, err := parseInterfaces([]byte(exampleJSON))
 	c.Check(err, jc.ErrorIsNil)
 	c.Check(result, jc.DeepEquals, expected)
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Legacy (pre 1.9) environs.NetworkInterfaces() implementation tests follow.
+
+const lshwXMLTemplate = `
+<?xml version="1.0" standalone="yes" ?>
+<!-- generated by lshw-B.02.16 -->
+<list>
+<node id="node1" claimed="true" class="system" handle="DMI:0001">
+ <description>Computer</description>
+ <product>VirtualBox ()</product>
+ <width units="bits">64</width>
+  <node id="core" claimed="true" class="bus" handle="DMI:0008">
+   <description>Motherboard</description>
+    <node id="pci" claimed="true" class="bridge" handle="PCIBUS:0000:00">
+     <description>Host bridge</description>{{$list := .}}{{range $mac, $ifi := $list}}
+      <node id="network{{if gt (len $list) 1}}:{{$ifi.DeviceIndex}}{{end}}"{{if $ifi.Disabled}} disabled="true"{{end}} claimed="true" class="network" handle="PCI:0000:00:03.0">
+       <description>Ethernet interface</description>
+       <product>82540EM Gigabit Ethernet Controller</product>
+       <logicalname>{{$ifi.InterfaceName}}</logicalname>
+       <serial>{{$mac}}</serial>
+      </node>{{end}}
+    </node>
+  </node>
+</node>
+</list>
+`
+
+func (suite *environSuite) generateHWTemplate(netMacs map[string]ifaceInfo) (string, error) {
+	tmpl, err := template.New("test").Parse(lshwXMLTemplate)
+	if err != nil {
+		return "", err
+	}
+	var buf bytes.Buffer
+	err = tmpl.Execute(&buf, netMacs)
+	if err != nil {
+		return "", err
+	}
+	return string(buf.Bytes()), nil
+}
+
+func (suite *environSuite) TestGetNetworkMACs(c *gc.C) {
+	env := suite.makeEnviron()
+
+	suite.testMAASObject.TestServer.NewNode(`{"system_id": "node_1"}`)
+	suite.testMAASObject.TestServer.NewNode(`{"system_id": "node_2"}`)
+	suite.testMAASObject.TestServer.NewNetwork(
+		`{"name": "net_1","ip":"0.1.2.0","netmask":"255.255.255.0"}`,
+	)
+	suite.testMAASObject.TestServer.NewNetwork(
+		`{"name": "net_2","ip":"0.2.2.0","netmask":"255.255.255.0"}`,
+	)
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node_2", "net_2", "aa:bb:cc:dd:ee:22")
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node_1", "net_1", "aa:bb:cc:dd:ee:11")
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node_2", "net_1", "aa:bb:cc:dd:ee:21")
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node_1", "net_2", "aa:bb:cc:dd:ee:12")
+
+	networks, err := env.getNetworkMACs("net_1")
+	c.Assert(err, jc.ErrorIsNil)
+	c.Check(networks, jc.SameContents, []string{"aa:bb:cc:dd:ee:11", "aa:bb:cc:dd:ee:21"})
+
+	networks, err = env.getNetworkMACs("net_2")
+	c.Assert(err, jc.ErrorIsNil)
+	c.Check(networks, jc.SameContents, []string{"aa:bb:cc:dd:ee:12", "aa:bb:cc:dd:ee:22"})
+
+	networks, err = env.getNetworkMACs("net_3")
+	c.Check(networks, gc.HasLen, 0)
+	c.Assert(err, jc.ErrorIsNil)
+}
+
+func (suite *environSuite) TestGetInstanceNetworks(c *gc.C) {
+	suite.getNetwork("test_network", 123, 321)
+	testInstance := suite.getInstance("instance_for_network")
+	suite.testMAASObject.TestServer.ConnectNodeToNetwork("instance_for_network", "test_network")
+	networks, err := suite.makeEnviron().getInstanceNetworks(testInstance)
+	c.Assert(err, jc.ErrorIsNil)
+	c.Check(networks, gc.DeepEquals, []networkDetails{
+		{Name: "test_network", IP: "192.168.123.1", Mask: "255.255.255.0", VLANTag: 321,
+			Description: "test_network_123_321"},
+	})
+}
+
+// A typical lshw XML dump with lots of things left out.
+const lshwXMLTestExtractInterfaces = `
+<?xml version="1.0" standalone="yes" ?>
+<!-- generated by lshw-B.02.16 -->
+<list>
+<node id="machine" claimed="true" class="system" handle="DMI:0001">
+ <description>Notebook</description>
+ <product>MyMachine</product>
+ <version>1.0</version>
+ <width units="bits">64</width>
+  <node id="core" claimed="true" class="bus" handle="DMI:0002">
+   <description>Motherboard</description>
+    <node id="cpu" claimed="true" class="processor" handle="DMI:0004">
+     <description>CPU</description>
+      <node id="pci:2" claimed="true" class="bridge" handle="PCIBUS:0000:03">
+        <node id="network:0" claimed="true" disabled="true" class="network" handle="PCI:0000:03:00.0">
+         <logicalname>wlan0</logicalname>
+         <serial>aa:bb:cc:dd:ee:ff</serial>
+        </node>
+        <node id="network:1" claimed="true" class="network" handle="PCI:0000:04:00.0">
+         <logicalname>eth0</logicalname>
+         <serial>aa:bb:cc:dd:ee:f1</serial>
+        </node>
+      </node>
+    </node>
+  </node>
+  <node id="network:2" claimed="true" class="network" handle="">
+   <logicalname>vnet1</logicalname>
+   <serial>aa:bb:cc:dd:ee:f2</serial>
+  </node>
+</node>
+</list>
+`
+
+// An lshw XML dump with implicit network interface indexes.
+const lshwXMLTestExtractInterfacesImplicitIndexes = `
+<?xml version="1.0" standalone="yes" ?>
+<!-- generated by lshw-B.02.16 -->
+<list>
+<node id="machine" claimed="true" class="system" handle="DMI:0001">
+ <description>Notebook</description>
+ <product>MyMachine</product>
+ <version>1.0</version>
+ <width units="bits">64</width>
+  <node id="core" claimed="true" class="bus" handle="DMI:0002">
+   <description>Motherboard</description>
+    <node id="cpu" claimed="true" class="processor" handle="DMI:0004">
+     <description>CPU</description>
+      <node id="pci:2" claimed="true" class="bridge" handle="PCIBUS:0000:03">
+        <node id="network" claimed="true" disabled="true" class="network" handle="PCI:0000:03:00.0">
+         <logicalname>wlan0</logicalname>
+         <serial>aa:bb:cc:dd:ee:ff</serial>
+        </node>
+        <node id="network" claimed="true" class="network" handle="PCI:0000:04:00.0">
+         <logicalname>eth0</logicalname>
+         <serial>aa:bb:cc:dd:ee:f1</serial>
+        </node>
+      </node>
+    </node>
+  </node>
+  <node id="network" claimed="true" class="network" handle="">
+   <logicalname>vnet1</logicalname>
+   <serial>aa:bb:cc:dd:ee:f2</serial>
+  </node>
+</node>
+</list>
+`
+
+func (suite *environSuite) TestExtractInterfaces(c *gc.C) {
+	rawData := []string{
+		lshwXMLTestExtractInterfaces,
+		lshwXMLTestExtractInterfacesImplicitIndexes,
+	}
+	for _, data := range rawData {
+		inst := suite.getInstance("testInstance")
+		interfaces, primaryIface, err := extractInterfaces(inst, []byte(data))
+		c.Assert(err, jc.ErrorIsNil)
+		c.Check(primaryIface, gc.Equals, "eth0")
+		c.Check(interfaces, jc.DeepEquals, map[string]ifaceInfo{
+			"aa:bb:cc:dd:ee:ff": {0, "wlan0", true},
+			"aa:bb:cc:dd:ee:f1": {1, "eth0", false},
+			"aa:bb:cc:dd:ee:f2": {2, "vnet1", false},
+		})
+	}
+}
+
+func (suite *environSuite) TestGetInstanceNetworkInterfaces(c *gc.C) {
+	inst := suite.getInstance("testInstance")
+	templateInterfaces := map[string]ifaceInfo{
+		"aa:bb:cc:dd:ee:ff": {0, "wlan0", true},
+		"aa:bb:cc:dd:ee:f1": {1, "eth0", true},
+		"aa:bb:cc:dd:ee:f2": {2, "vnet1", false},
+	}
+	lshwXML, err := suite.generateHWTemplate(templateInterfaces)
+	c.Assert(err, jc.ErrorIsNil)
+
+	suite.testMAASObject.TestServer.AddNodeDetails("testInstance", lshwXML)
+	env := suite.makeEnviron()
+	interfaces, primaryIface, err := env.getInstanceNetworkInterfaces(inst)
+	c.Assert(err, jc.ErrorIsNil)
+	// Both wlan0 and eth0 are disabled in lshw output.
+	c.Check(primaryIface, gc.Equals, "vnet1")
+	c.Check(interfaces, jc.DeepEquals, templateInterfaces)
+}
+
+func (suite *environSuite) TestSetupNetworks(c *gc.C) {
+	testInstance := suite.getInstance("node1")
+	templateInterfaces := map[string]ifaceInfo{
+		"aa:bb:cc:dd:ee:ff": {0, "wlan0", true},
+		"aa:bb:cc:dd:ee:f1": {1, "eth0", true},
+		"aa:bb:cc:dd:ee:f2": {2, "vnet1", false},
+	}
+	lshwXML, err := suite.generateHWTemplate(templateInterfaces)
+	c.Assert(err, jc.ErrorIsNil)
+
+	suite.testMAASObject.TestServer.AddNodeDetails("node1", lshwXML)
+	suite.getNetwork("LAN", 2, 42)
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node1", "LAN", "aa:bb:cc:dd:ee:f1")
+	suite.getNetwork("Virt", 3, 0)
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node1", "Virt", "aa:bb:cc:dd:ee:f2")
+	suite.getNetwork("WLAN", 1, 0)
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node1", "WLAN", "aa:bb:cc:dd:ee:ff")
+	networkInfo, primaryIface, err := suite.makeEnviron().setupNetworks(
+		testInstance,
+		set.NewStrings("WLAN"), // Disable WLAN only.
+	)
+	c.Assert(err, jc.ErrorIsNil)
+
+	// Note: order of networks is based on lshwXML
+	c.Check(primaryIface, gc.Equals, "vnet1")
+	// Unfortunately, because network.InterfaceInfo is unhashable
+	// (contains a map) we can't use jc.SameContents here.
+	c.Check(networkInfo, gc.HasLen, 3)
+	for _, info := range networkInfo {
+		switch info.DeviceIndex {
+		case 0:
+			c.Check(info, jc.DeepEquals, network.InterfaceInfo{
+				MACAddress:    "aa:bb:cc:dd:ee:ff",
+				CIDR:          "192.168.1.1/24",
+				NetworkName:   "WLAN",
+				ProviderId:    "WLAN",
+				VLANTag:       0,
+				DeviceIndex:   0,
+				InterfaceName: "wlan0",
+				Disabled:      true, // from networksToDisable("WLAN")
+			})
+		case 1:
+			c.Check(info, jc.DeepEquals, network.InterfaceInfo{
+				DeviceIndex:   1,
+				MACAddress:    "aa:bb:cc:dd:ee:f1",
+				CIDR:          "192.168.2.1/24",
+				NetworkName:   "LAN",
+				ProviderId:    "LAN",
+				VLANTag:       42,
+				InterfaceName: "eth0",
+				Disabled:      true, // from networksToDisable("WLAN")
+			})
+		case 2:
+			c.Check(info, jc.DeepEquals, network.InterfaceInfo{
+				MACAddress:    "aa:bb:cc:dd:ee:f2",
+				CIDR:          "192.168.3.1/24",
+				NetworkName:   "Virt",
+				ProviderId:    "Virt",
+				VLANTag:       0,
+				DeviceIndex:   2,
+				InterfaceName: "vnet1",
+				Disabled:      false,
+			})
+		}
+	}
+}
+
+// The same test, but now "Virt" network does not have matched MAC address
+func (suite *environSuite) TestSetupNetworksPartialMatch(c *gc.C) {
+	testInstance := suite.getInstance("node1")
+	templateInterfaces := map[string]ifaceInfo{
+		"aa:bb:cc:dd:ee:ff": {0, "wlan0", true},
+		"aa:bb:cc:dd:ee:f1": {1, "eth0", false},
+		"aa:bb:cc:dd:ee:f2": {2, "vnet1", false},
+	}
+	lshwXML, err := suite.generateHWTemplate(templateInterfaces)
+	c.Assert(err, jc.ErrorIsNil)
+
+	suite.testMAASObject.TestServer.AddNodeDetails("node1", lshwXML)
+	suite.getNetwork("LAN", 2, 42)
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node1", "LAN", "aa:bb:cc:dd:ee:f1")
+	suite.getNetwork("Virt", 3, 0)
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node1", "Virt", "aa:bb:cc:dd:ee:f3")
+	networkInfo, primaryIface, err := suite.makeEnviron().setupNetworks(
+		testInstance,
+		set.NewStrings(), // All enabled.
+	)
+	c.Assert(err, jc.ErrorIsNil)
+
+	// Note: order of networks is based on lshwXML
+	c.Check(primaryIface, gc.Equals, "eth0")
+	c.Check(networkInfo, jc.DeepEquals, []network.InterfaceInfo{{
+		MACAddress:    "aa:bb:cc:dd:ee:f1",
+		CIDR:          "192.168.2.1/24",
+		NetworkName:   "LAN",
+		ProviderId:    "LAN",
+		VLANTag:       42,
+		DeviceIndex:   1,
+		InterfaceName: "eth0",
+		Disabled:      false,
+	}})
+}
+
+// The same test, but now no networks have matched MAC
+func (suite *environSuite) TestSetupNetworksNoMatch(c *gc.C) {
+	testInstance := suite.getInstance("node1")
+	templateInterfaces := map[string]ifaceInfo{
+		"aa:bb:cc:dd:ee:ff": {0, "wlan0", true},
+		"aa:bb:cc:dd:ee:f1": {1, "eth0", false},
+		"aa:bb:cc:dd:ee:f2": {2, "vnet1", false},
+	}
+	lshwXML, err := suite.generateHWTemplate(templateInterfaces)
+	c.Assert(err, jc.ErrorIsNil)
+
+	suite.testMAASObject.TestServer.AddNodeDetails("node1", lshwXML)
+	suite.getNetwork("Virt", 3, 0)
+	suite.testMAASObject.TestServer.ConnectNodeToNetworkWithMACAddress("node1", "Virt", "aa:bb:cc:dd:ee:f3")
+	networkInfo, primaryIface, err := suite.makeEnviron().setupNetworks(
+		testInstance,
+		set.NewStrings(), // All enabled.
+	)
+	c.Assert(err, jc.ErrorIsNil)
+
+	// Note: order of networks is based on lshwXML
+	c.Check(primaryIface, gc.Equals, "eth0")
+	c.Check(networkInfo, gc.HasLen, 0)
+}
+
+func (suite *environSuite) TestNetworkInterfacesLegacy(c *gc.C) {
+	testInstance := suite.createSubnets(c, false)
+
+	netInfo, err := suite.makeEnviron().NetworkInterfaces(testInstance.Id())
+	c.Assert(err, jc.ErrorIsNil)
+
+	expectedInfo := []network.InterfaceInfo{{
+		DeviceIndex:      0,
+		MACAddress:       "aa:bb:cc:dd:ee:ff",
+		CIDR:             "192.168.1.1/24",
+		ProviderSubnetId: "WLAN",
+		VLANTag:          0,
+		InterfaceName:    "wlan0",
+		Disabled:         true,
+		NoAutoStart:      true,
+		ConfigType:       network.ConfigDHCP,
+		ExtraConfig:      nil,
+		GatewayAddress:   network.Address{},
+		Address:          network.NewScopedAddress("192.168.1.1", network.ScopeCloudLocal),
+	}, {
+		DeviceIndex:      1,
+		MACAddress:       "aa:bb:cc:dd:ee:f1",
+		CIDR:             "192.168.2.1/24",
+		ProviderSubnetId: "LAN",
+		VLANTag:          42,
+		InterfaceName:    "eth0",
+		Disabled:         false,
+		NoAutoStart:      false,
+		ConfigType:       network.ConfigDHCP,
+		ExtraConfig:      nil,
+		GatewayAddress:   network.Address{},
+		Address:          network.NewScopedAddress("192.168.2.1", network.ScopeCloudLocal),
+	}, {
+		DeviceIndex:      2,
+		MACAddress:       "aa:bb:cc:dd:ee:f2",
+		CIDR:             "192.168.3.1/24",
+		ProviderSubnetId: "Virt",
+		VLANTag:          0,
+		InterfaceName:    "vnet1",
+		Disabled:         false,
+		NoAutoStart:      false,
+		ConfigType:       network.ConfigDHCP,
+		ExtraConfig:      nil,
+		GatewayAddress:   network.Address{},
+		Address:          network.NewScopedAddress("192.168.3.1", network.ScopeCloudLocal),
+	}}
+	network.SortInterfaceInfo(netInfo)
+	c.Assert(netInfo, jc.DeepEquals, expectedInfo)
 }
